+++ bin/convert_markdown_images_to_captions.py
@@ -1,4 +1,8 @@
 #!/usr/bin/env python3
+"""
+Convert markdown image links in an RMarkdown document to use the nujucks image macros
+from the `_include` directory.
+"""
 
 import re
 import textwrap
@@ -6,9 +10,9 @@ import textwrap
 import click
 
 
-def convert_link(document: str, src_image_dir: str, dst_image_url: str) -> str:
+def convert_links(document: str, src_image_dir: str, dst_image_url: str) -> str:
 
-    matches = re.finditer(r'^!\[(.*?)\]\((.+?)\)$', document, flags=re.MULTILINE | re.DOTALL)
+    matches = re.finditer(r"^!\[(.*?)\]\((.+?)\)$", document, flags=re.MULTILINE | re.DOTALL)
     for match in matches:
         description, link = match.group(1), match.group(2)
 
@@ -23,42 +27,46 @@ def convert_link(document: str, src_image_dir: str, dst_image_url: str) -> str:
 
         # If there are multiple sentences in the description, then make the first sentence the bolded text
         # and the rest the longer description.
-        fields = description.strip().split('.', 1)
+        fields = description.strip().split(".", 1)
 
-        args = {
-            "url": url,
-            "anchor": '',
-            "short_desc": fields[0].strip().replace("\n", " ") + "."
-        }
+        args = {"url": url, "anchor": "", "short_desc": fields[0].strip().replace("\n", " ") + "."}
         if fields[1]:
             args["long_desc"] = fields[1].strip().replace("\n", " ")
 
-        arg_str =  ', '.join(f"{k}: '{v}'" for k, v in args.items())
+        arg_str = ", ".join(f"{k}: '{v}'" for k, v in args.items())
 
         # Create the new link
-        nunjucks_link = textwrap.dedent(f"""\
-        {{% include 'image_with_caption.njk', 
+        nunjucks_link = textwrap.dedent(
+            f"""\
+        {{% include 'image_with_caption.njk',
             {arg_str}
-        %}}""")
+        %}}"""
+        )
 
         document = document.replace(match.group(0), nunjucks_link)
     return document
 
 
-
-
 @click.command("Convert RMarkdown output to nujucks markdown.")
-@click.argument("input_file", type=click.Path(exists=True, file_okay=True, dir_okay=False), required=True)
-@click.argument("output_file", type=click.Path(exists=False, file_okay=True, dir_okay=False), required=True)
-# TODO: Change back exists=False
-@click.option("--src-image-dir", type=click.Path(exists=False, file_okay=False, dir_okay=True), required=True)
+@click.argument(
+    "input_file", type=click.Path(exists=True, file_okay=True, dir_okay=False), required=True
+)
+@click.argument(
+    "output_file", type=click.Path(exists=False, file_okay=True, dir_okay=False), required=True
+)
+@click.option(
+    "--src-image-dir", type=click.Path(exists=False, file_okay=False, dir_okay=True), required=True
+)
 @click.option("--dst-image-url", type=str, required=True)
 @click.option("--lede", is_flag=True)
-def main(input_file: str, output_file: str, src_image_dir: str, dst_image_url: str, lede: bool = False) -> None:
+def main(
+    input_file: str, output_file: str, src_image_dir: str, dst_image_url: str, lede: bool = False
+) -> None:
     with open(input_file, "rt") as fh_in:
-        transform = convert_link(fh_in.read(), src_image_dir, dst_image_url)
+        transform = convert_links(fh_in.read(), src_image_dir, dst_image_url)
 
-    # If this article has a lede, then add the lede markdown container
+    # If the `--lede` flag is given, then add the lede markdown container to the
+    # first paragraph.
     if lede:
         paragraphs = transform.split("\n\n")
         paragraphs[0] = f":::lede\n{paragraphs[0]}\n:::\n"